% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bulk.R
\name{bulk_utils_find_regulators_with_recurring_edges}
\alias{bulk_utils_find_regulators_with_recurring_edges}
\title{Find recurring regulators}
\usage{
bulk_utils_find_regulators_with_recurring_edges(weightMatList, plotConnections)
}
\arguments{
\item{weightMatList}{a list of (weighted) adjacency matrices; 
each list element must be an adjacency matrix with regulators in rows,
targets in columns}

\item{plotConnections}{the number of connections to subset to}
}
\value{
A vector containing the names of the recurring regulators
}
\description{
This function finds regulators that appear as the 
same network edge in more than one of the input networks.
}
\keyword{internal}
