% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bulk.R
\name{bulk_utils_get_link_list_rename}
\alias{bulk_utils_get_link_list_rename}
\title{Convert the adjacency matrix to network links}
\usage{
bulk_utils_get_link_list_rename(weightMat, plotConnections)
}
\arguments{
\item{weightMat}{the (weighted) adjacency matrix - regulators in rows,
targets in columns}

\item{plotConnections}{the number of connections to subset to}
}
\value{
A data frame with fields from, to and value, describing the edges
of the network
}
\description{
This function converts an adjacency matrix to a data frame
of network links, subset to the most important ones.
}
\keyword{internal}
