% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bulk.R
\name{bulk_utils_pca_contrib}
\alias{bulk_utils_pca_contrib}
\title{Plot PCA Contributions}
\usage{
bulk_utils_pca_contrib(pca.res, comp = 1, anno)
}
\arguments{
\item{pca.res}{Result of prcomp or similar PCA function.}

\item{comp}{Integer; which component to plot (default: 1).}

\item{anno}{Data frame with annotation, must include 'display_name' and 'm_z'.}
}
\value{
List with elements: 'plot' (ggplot object) and 'table' (data frame of top contributors).
}
\description{
Plots the top contributing features to a selected PCA component.
}
\keyword{internal}
