% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bulk.R
\name{bulk_utils_perform_plsda}
\alias{bulk_utils_perform_plsda}
\title{Perform PLS-DA}
\usage{
bulk_utils_perform_plsda(intensity.matrix, metadata, separator.id)
}
\arguments{
\item{intensity.matrix}{Numeric matrix of intensities (features × samples).}

\item{metadata}{Data frame of sample metadata.}

\item{separator.id}{Integer index of metadata column to use as group/class.}
}
\value{
PLS-DA result object from mixOmics::plsda.
}
\description{
Runs PLS-DA (Partial Least Squares Discriminant Analysis) on intensity matrix and metadata.
}
\keyword{internal}
