% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bulk.R
\name{bulk_utils_plot_GRN}
\alias{bulk_utils_plot_GRN}
\title{Plot a GRN}
\usage{
bulk_utils_plot_GRN(
  weightMat,
  anno,
  plotConnections,
  plot_position_grid,
  n_networks,
  recurring_regulators
)
}
\arguments{
\item{weightMat}{the (weighted) adjacency matrix - regulators in rows,
targets in columns}

\item{anno}{a data frame with peak annotation including 'display_name' (peak names)
and 'm_z' (peak IDs); used to highlight target peaks in the network}

\item{plotConnections}{the number of connections to subset to for plotting}

\item{plot_position_grid, n_networks}{the position of the plot in 
the grid (1-4) and the number of networks shown (1-4); these are
solely used for hiding unwanted plots in the shiny app}

\item{recurring_regulators}{targets to be highlighted; usually the
result of \code{\link{bulk_utils_find_regulators_with_recurring_edges}}}
}
\value{
A network plot. See visNetwork package for more details.
}
\description{
This function creates a network plot of a GRN.
}
\keyword{internal}
