% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bulk.R
\name{bulk_utils_plot_upset}
\alias{bulk_utils_plot_upset}
\title{Visualise the overlap of edges between different networks}
\usage{
bulk_utils_plot_upset(weightMatList, plotConnections)
}
\arguments{
\item{weightMatList}{a list of (weighted) adjacency matrices; 
each list element must be an adjacency matrix with regulators in rows,
targets in columns}

\item{plotConnections}{the number of connections to subset to}
}
\value{
An UpSet plot. See UpSetR package for more details.
}
\description{
This function creates an UpSet plot of the intersections
and specific differences of the edges in the input networks.
}
\keyword{internal}
