% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPanel_DATab.R
\name{BulkDEUI}
\alias{BulkDEUI}
\title{Bulk Differential Expression Panel UI}
\usage{
BulkDEUI(id, bulk.metadata, show = TRUE)
}
\arguments{
\item{id}{Module namespace (character).}

\item{show}{Logical; whether to render the panel (default: TRUE).}

\item{bulk_metadata}{Data frame with bulk sample metadata.}
}
\value{
A [shiny::tabPanel] containing the UI elements.
}
\description{
Creates the user interface for differential intensity analysis.
}
\keyword{internal}
