% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPanel_MultiCompareTab.R
\name{bulkPanel_MultiCompareTab}
\alias{bulkPanel_MultiCompareTab}
\alias{BulkMultiCompareUI}
\alias{BulkMultiCompareServer}
\title{Bulk Multi-Comparison UI/Server}
\usage{
BulkMultiCompareUI(id, bulk.metadata, bulk.intensity.matrix)

BulkMultiCompareServer(id, bulk.intensity.matrix, bulk.metadata, anno)
}
\description{
UI and server for multi-group comparisons in the bulk dataset using ANOVA/Kruskal-Wallis and post-hoc tests, with boxplots, cross-plots, and Venn diagrams.
}
