% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing_autocorrelation.R
\name{compute_autocorrelation}
\alias{compute_autocorrelation}
\title{Compute spatial autocorrelation (correlation between feature and spatial lag)}
\usage{
compute_autocorrelation(intensity, lag_mat)
}
\arguments{
\item{intensity}{Matrix (spots x features).}

\item{lag_mat}{Matrix (spots x features).}
}
\value{
Named vector of autocorrelation per feature.
}
\description{
Compute spatial autocorrelation (correlation between feature and spatial lag)
}
\keyword{autocorrelation}
\keyword{spatial,}
