% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing_autocorrelation.R
\name{compute_spatial_lag}
\alias{compute_spatial_lag}
\title{Compute spatial lag for each spot}
\usage{
compute_spatial_lag(intensity, neighbour_list)
}
\arguments{
\item{intensity}{Matrix (spots x features), rownames = spot_id.}

\item{neighbour_list}{List of neighbour spot_ids per spot.}
}
\value{
Matrix of spatial lags (spots x features).
}
\description{
Compute spatial lag for each spot
}
\keyword{lag}
\keyword{spatial,}
