% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_DATab.R
\name{create_bulk_exp_regions}
\alias{create_bulk_exp_regions}
\title{Region differential expression utilities}
\usage{
create_bulk_exp_regions(
  intensity.matrix,
  metadata,
  bulk.metadata,
  sample.id.metadata.column,
  region.ids,
  minimum.pixels = 3
)
}
\description{
Functions to create pseudobulk representations of user-defined regions and prepare metadata for region-level DE.
}
\keyword{internal}
