% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{GCD}
\alias{GCD}
\title{Greatest common divisor for grid spacing}
\usage{
GCD(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Integer; greatest common divisor.
}
\description{
Computes the greatest common divisor for a numeric vector, used for grid spacing.
}
\keyword{gcd}
\keyword{grid,}
