% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPanel_GRNFuns.R
\name{GRNFuns}
\alias{GRNFuns}
\title{Helper functions for Metabolite Peak Regulatory Network (GRN) Analysis}
\description{
These functions support the GRN inference and visualization modules.
}
\keyword{internal}
