% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPanel_GRNTab.R, R/regionPanel_GRNTab.R
\name{GRNpanel}
\alias{GRNpanel}
\alias{BulkGRNUI}
\alias{BulkGRNServer}
\alias{RegionGRNUI}
\alias{RegionGRNServer}
\title{Metabolite Peak Regulatory Network (GRN) UI and Server Module}
\usage{
BulkGRNUI(id, bulk.metadata, show = TRUE)

BulkGRNServer(id, bulk.intensity.matrix, bulk.metadata, anno)

RegionGRNUI(id, metadata, show = TRUE)

RegionGRNServer(
  id,
  intensity.matrix,
  metadata,
  anno,
  bulk.metadata,
  shared_data
)
}
\value{
The UI and Server components of the shiny module, that can be used
within the UI and Server definitions of a shiny app.
}
\description{
This file contains the UI and server logic for the GRN panel. Helper functions for GRN inference and network processing are now located in bulkPanel_GRNFuns.R.

These are the UI and server components of the GRN panel of the 
shiny app. It is generated by including 'GRN' in the panels.default argument
of \code{generateShinyApp}.
}
\keyword{internal}
