% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{info_dropmenu}
\alias{info_dropmenu}
\title{Convenience wrapper for a consistent info drop menu}
\usage{
info_dropmenu(ns, id, title, items)
}
\arguments{
\item{ns}{Namespace function.}

\item{id}{Character; base id for the info button.}

\item{title}{Character; title string for the menu.}

\item{items}{Character vector of bullet items.}
}
\description{
Creates a consistent info drop menu using shinyWidgets::dropMenu.
}
\keyword{info,}
\keyword{menu}
\keyword{shiny,}
