% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introPanel_Overview.R
\name{IntroOverviewServer}
\alias{IntroOverviewServer}
\title{SMEW App Overview Server}
\usage{
IntroOverviewServer(id)
}
\arguments{
\item{id}{Shiny module id}
}
\value{
A shiny tabPanel object for the overview tab
}
\description{
Server logic for the SMEW app overview tab, with placeholders for manuscript and GitHub links.
}
