% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introPanel_Overview.R
\name{IntroOverviewUI}
\alias{IntroOverviewUI}
\title{SMEW App Overview UI}
\usage{
IntroOverviewUI(id)
}
\arguments{
\item{id}{Shiny module id}
}
\value{
A shiny tabPanel object for the overview tab
}
\description{
UI for the SMEW app overview tab, with placeholders for manuscript and GitHub links.
}
