% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introPanel_SpatialVis.R
\name{IntroSpatialVisServer}
\alias{IntroSpatialVisServer}
\title{Spatial Visualisation Panel Server}
\usage{
IntroSpatialVisServer(
  id,
  bulk.metadata,
  full.metadata,
  full.intensity.matrix,
  anno
)
}
\arguments{
\item{id}{Shiny module id}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{full.metadata}{Data frame of full sample metadata}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{anno}{Data frame of peak annotations}
}
\value{
None; called for side effects in Shiny module
}
\description{
Provides the server logic for spatial visualisation of peaks and metadata in the SMEW app.
}
