% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introPanel_SpatialVis.R
\name{IntroSpatialVisUI}
\alias{IntroSpatialVisUI}
\title{Spatial Visualisation Panel UI}
\usage{
IntroSpatialVisUI(id, bulk.metadata, full.metadata, show = TRUE)
}
\arguments{
\item{id}{Shiny module id}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{full.metadata}{Data frame of full sample metadata}

\item{show}{Logical; whether to show the panel (default TRUE)}
}
\value{
A shiny tabPanel object for spatial visualisation
}
\description{
Provides the UI for spatial visualisation of peaks and metadata in the SMEW app.
}
