% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_ComparisonTab.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Jaccard Similarity Index}
\usage{
jaccard_index(x, y)
}
\arguments{
\item{x}{Vector of cluster assignments (first method)}

\item{y}{Vector of cluster assignments (second method)}
}
\value{
Numeric value between 0 and 1 representing Jaccard index
}
\description{
Calculate Jaccard similarity between two vectors
}
\keyword{internal}
