% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{map_peaks_to_kegg}
\alias{map_peaks_to_kegg}
\title{Map peak masses to KEGG names}
\usage{
map_peaks_to_kegg(
  peak_list,
  kegg_db,
  adducts,
  ion_mode,
  neg_adduct_formulas,
  pos_adduct_formulas,
  ppm = 5
)
}
\arguments{
\item{peak_list}{Numeric vector of experimental m/z values.}

\item{kegg_db}{Data frame with KEGG compound information.}

\item{adducts}{Character vector of adduct names to use.}

\item{ion_mode}{Character; ionization mode - either "Positive" or "Negative".}

\item{neg_adduct_formulas, pos_adduct_formulas}{Data frames with negative/positive adduct formulas.}

\item{ppm}{Numeric; mass tolerance in ppm (default: 5).}
}
\value{
Data frame mapping each peak to KEGG compound/adduct info.
}
\description{
Maps a vector of m/z peak values to KEGG compound names and adducts using get_matched_peaks.
}
\keyword{KEGG}
\keyword{annotation,}
