% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelPanel_Enrichment.R
\name{PixelEnrichmentServer}
\alias{PixelEnrichmentServer}
\title{Pixel Enrichment Server}
\usage{
PixelEnrichmentServer(
  id,
  bulk.metadata,
  full.metadata,
  full.intensity.matrix,
  anno,
  pixel_enrichment
)
}
\arguments{
\item{id}{Shiny module id}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{full.metadata}{Data frame of full sample metadata}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{anno}{Data frame of peak annotations}

\item{pixel_enrichment}{List or data structure with pixel-level enrichment results}
}
\value{
None; called for side effects in Shiny module
}
\description{
PixelEnrichmentServer
}
\details{
Provides the server logic for spatial pathway enrichment analysis in the SMEW app.
}
