% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelPanel_SVM.R
\name{PixelSVMServer}
\alias{PixelSVMServer}
\title{Pixel SVM Server}
\usage{
PixelSVMServer(
  id,
  bulk.metadata,
  full.metadata,
  full.intensity.matrix,
  anno,
  DEresults,
  svm_identification,
  spatial.cross.cor
)
}
\arguments{
\item{id}{Shiny module id}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{full.metadata}{Data frame of full sample metadata}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{anno}{Data frame of peak annotations}

\item{DEresults}{Reactive expression or function returning DE results}

\item{svm_identification}{List of SVM identification results per sample}

\item{spatial.cross.cor}{Matrix of spatial cross-correlation values}
}
\value{
None; called for side effects in Shiny module
}
\description{
Provides the server logic for spatially variable peak analysis using SVM in the SMEW app.
}
