% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPanel_QCPlotFuns.R
\name{plot_pca}
\alias{plot_pca}
\title{QC plotting helpers}
\usage{
plot_pca(
  pca.res,
  metadata,
  annotation.id,
  show.confidence.ellipses = TRUE,
  label.force = 1
)
}
\description{
Utilities for plotting PCA/PLS-DA and other QC plots used in the Bulk QC panel.
}
\keyword{internal}
