% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateShinyApp.R
\name{preprocess_for_app}
\alias{preprocess_for_app}
\title{Preprocess Data for App Generation}
\usage{
preprocess_for_app(
  intensity_csv,
  metadata_csv,
  output_dir,
  denoise = F,
  anno = NULL,
  adducts = NULL,
  ion_mode = NULL,
  organism = "Human",
  ppm = 10,
  only_annotated = FALSE,
  histology_images_dir = NULL,
  run_autocorrelation = FALSE,
  top_autocorrelated_peaks = 10,
  n_cores = 1,
  run_pixel_enrichment = FALSE,
  multi_modal_path = NULL
)
}
\arguments{
\item{intensity_csv}{Path to the intensity matrix CSV file}

\item{metadata_csv}{Path to the metadata CSV file}

\item{output_dir}{Directory to save processed data}

\item{denoise}{Logical; whether to denoise the data (default FALSE)}

\item{anno}{Optional annotation data.frame or NULL}

\item{adducts}{Optional adducts table or NULL}

\item{ion_mode}{Ionization mode (character or NULL)}

\item{organism}{Organism name (default 'Human')}

\item{ppm}{Numeric; mass accuracy in ppm (default 10)}

\item{only_annotated}{Logical; if TRUE, only annotated peaks are used (default FALSE)}

\item{histology_images_dir}{Optional directory with histology images}

\item{run_autocorrelation}{Logical; whether to run spatial autocorrelation (default FALSE)}

\item{top_autocorrelated_peaks}{Integer; number of top autocorrelated peaks to use (default 10)}

\item{n_cores}{Number of cores for parallel processing (default 1)}

\item{run_pixel_enrichment}{Logical; whether to run pixel-level enrichment (default FALSE)}
}
\value{
List or path to processed data
}
\description{
This function runs all necessary preprocessing steps on the input data for the SMEW app.
}
