% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{preprocessing_calculate_gcd}
\alias{preprocessing_calculate_gcd}
\title{Greatest common divisor for grid spacing}
\usage{
preprocessing_calculate_gcd(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Integer; greatest common divisor.
}
\description{
Computes the greatest common divisor for a numeric vector, used for grid spacing.
}
\keyword{internal}
