% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{preprocessing_pick.neighbours.per.pixel}
\alias{preprocessing_pick.neighbours.per.pixel}
\title{Denoise a single pixel by neighbourhood}
\usage{
preprocessing_pick.neighbours.per.pixel(
  i,
  cor.matrix,
  distance.matrix,
  current.intensity,
  gcd,
  w
)
}
\arguments{
\item{i}{Pixel identifier (row name).}

\item{cor.matrix}{Correlation matrix between pixels.}

\item{distance.matrix}{Spatial distance matrix between pixels.}

\item{current.intensity}{Intensity matrix for current sample.}

\item{gcd}{Numeric; grid spacing value for the sample.}

\item{w}{Weight matrix (correlation * distance).}
}
\value{
Numeric vector of denoised intensities for pixel i.
}
\description{
Computes denoised intensity for a pixel by weighted averaging of its most correlated spatial neighbours.
}
\keyword{internal}
