% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing_autocorrelation.R
\name{preprocessing_spatial_autocorrelation_pipeline}
\alias{preprocessing_spatial_autocorrelation_pipeline}
\title{Main pipeline for spatial autocorrelation}
\usage{
preprocessing_spatial_autocorrelation_pipeline(intensity, metadata)
}
\arguments{
\item{intensity}{Matrix (pixels x features), rownames = pixel_id.}

\item{metadata}{Data frame with pixel_id, x_tf, y_tf, Sample.}
}
\value{
Named list of autocorrelation vectors per Sample.
}
\description{
Main pipeline for spatial autocorrelation
}
\keyword{internal}
