% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_region.R
\name{region_utils_colour_blender}
\alias{region_utils_colour_blender}
\title{Colour blender}
\usage{
region_utils_colour_blender(data, channels.use = NULL)
}
\arguments{
\item{data}{Matrix or data frame of numeric values (n x 2 or n x 3).}

\item{channels.use}{Character vector of length 2 or 3 specifying which colour channel to assign to each column of data. Must be a permutation of 'red', 'green', 'blue'.}
}
\value{
Character vector of RGB colour codes (length n).
}
\description{
Blend numeric feature/dimension values into RGB colour codes for visualisation.
}
\details{
Given a matrix or data frame of 2 or 3 numeric features/dimensions, and a vector specifying which colour channel (red, green, blue) to assign to each, this function returns RGB colour codes for each row. Used for coloured plotting in region analysis.


If data has 2 columns, the third channel is set to zero. Checks for valid and unique channel assignments. Used for coloured plotting in region analysis.
}
\keyword{internal}
