% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_ComparisonTab.R
\name{RegionComparisonServer}
\alias{RegionComparisonServer}
\title{RegionComparisonServer}
\usage{
RegionComparisonServer(id, shared_data, bulk.metadata)
}
\arguments{
\item{id}{Shiny module id}

\item{shared_data}{Reactive or shared data object}

\item{bulk.metadata}{Data frame of bulk sample metadata}
}
\value{
None; called for side effects in Shiny module
}
\description{
Server logic for the cross-clustering comparison tab.
}
