% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_ComparisonTab.R
\name{RegionComparisonUI}
\alias{RegionComparisonUI}
\title{RegionComparisonUI}
\usage{
RegionComparisonUI(id)
}
\arguments{
\item{id}{Shiny module id}
}
\value{
A shiny tabPanel object for the cross-clustering comparison tab
}
\description{
UI for the cross-clustering comparison tab.
}
