% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_HistologyTab.R
\name{RegionHistologyServer}
\alias{RegionHistologyServer}
\title{RegionHistologyServer}
\usage{
RegionHistologyServer(
  id,
  full.intensity.matrix,
  full.metadata,
  bulk.metadata,
  anno,
  shared_data
)
}
\arguments{
\item{id}{Shiny module id}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{full.metadata}{Data frame of full sample metadata}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{anno}{Data frame of peak annotations}

\item{shared_data}{Reactive or shared data object}
}
\value{
None; called for side effects in Shiny module
}
\description{
Server logic for manual region selection and histology overlay tab.
}
