% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_HistologyTab.R
\name{RegionHistologyUI}
\alias{RegionHistologyUI}
\title{RegionHistologyUI}
\usage{
RegionHistologyUI(
  id,
  bulk.metadata,
  full.metadata,
  full.intensity.matrix,
  show = TRUE
)
}
\arguments{
\item{id}{Shiny module id}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{full.metadata}{Data frame of full sample metadata}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{show}{Logical; whether to show the panel (default TRUE)}
}
\value{
A shiny tabPanel object for the manual region selection tab
}
\description{
UI for manual region selection and histology overlay tab.
}
