% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_RadialDistanceTab.R
\name{RegionRadialDistanceServer}
\alias{RegionRadialDistanceServer}
\title{RegionRadialDistanceServer}
\usage{
RegionRadialDistanceServer(
  id,
  full.intensity.matrix,
  full.metadata,
  bulk.metadata,
  shared_data,
  anno
)
}
\arguments{
\item{id}{Shiny module id}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{full.metadata}{Data frame of full sample metadata}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{shared_data}{Reactive or shared data object}

\item{anno}{Data frame of peak annotations}
}
\value{
None; called for side effects in Shiny module
}
\description{
Server logic for the region-level radial distance correlation tab in the SMEW app.
}
