% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_RadialDistanceTab.R
\name{RegionRadialDistanceUI}
\alias{RegionRadialDistanceUI}
\title{RegionRadialDistanceUI}
\usage{
RegionRadialDistanceUI(id, bulk.metadata)
}
\arguments{
\item{id}{Shiny module id}

\item{bulk.metadata}{Data frame of bulk sample metadata}
}
\value{
A shiny tabPanel object for the radial distance correlation tab
}
\description{
UI for the region-level radial distance correlation tab in the SMEW app.
}
