% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_SpatialSmoothingTab.R
\name{RegionSmoothingServer}
\alias{RegionSmoothingServer}
\title{RegionSmoothingServer}
\usage{
RegionSmoothingServer(id, shared_data, full.metadata)
}
\arguments{
\item{id}{Shiny module id}

\item{shared_data}{Reactive or shared data object}

\item{full.metadata}{Data frame of full sample metadata}
}
\value{
None; called for side effects in Shiny module
}
\description{
Server logic for the region smoothing tab.
}
