% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_SpatialSmoothingTab.R
\name{RegionSmoothingUI}
\alias{RegionSmoothingUI}
\title{RegionSmoothingUI}
\usage{
RegionSmoothingUI(id)
}
\arguments{
\item{id}{Shiny module id}
}
\value{
A shiny tabPanel object for the region smoothing tab
}
\description{
UI for the region smoothing tab.
}
