% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_SpatialClusters.R
\name{RegionSpatialClusterPanelServer}
\alias{RegionSpatialClusterPanelServer}
\title{RegionSpatialClusterPanelServer}
\usage{
RegionSpatialClusterPanelServer(
  id,
  full.intensity.matrix,
  full.metadata,
  bulk.metadata,
  anno = NULL,
  shared_data = NULL
)
}
\arguments{
\item{id}{Shiny module id}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{full.metadata}{Data frame of full sample metadata}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{anno}{Data frame of peak annotations}

\item{shared_data}{Reactive or shared data object}
}
\value{
None; called for side effects in Shiny module
}
\description{
Server logic for the region-level spatial clustering tab in the SMEW app.
}
