% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_VotingTab.R
\name{RegionVotingServer}
\alias{RegionVotingServer}
\title{RegionVotingServer}
\usage{
RegionVotingServer(
  id,
  bulk.metadata,
  full.metadata,
  full.intensity.matrix,
  anno,
  shared_data
)
}
\arguments{
\item{id}{Shiny module id}

\item{bulk.metadata}{Data frame of bulk sample metadata}

\item{full.metadata}{Data frame of full sample metadata}

\item{full.intensity.matrix}{Matrix of intensities (features x samples)}

\item{anno}{Data frame of peak annotations}

\item{shared_data}{Reactive or shared data object}
}
\value{
None; called for side effects in Shiny module
}
\description{
Server logic for the region-level voting scheme tab in the SMEW app.
}
