% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPanel_SpatialSmoothingTab.R
\name{smoothed.cluster}
\alias{smoothed.cluster}
\title{Spatial smoothing helpers}
\usage{
smoothed.cluster(metadata.sub, cluster.col)
}
\description{
Functions for smoothing cluster labels across spatial neighborhoods (used by the Cluster Smoothing tab).
}
\keyword{internal}
