% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateShinyApp.R
\name{write_static_app_r}
\alias{write_static_app_r}
\title{Write Example app.R Script}
\usage{
write_static_app_r(
  app_dir,
  run_enrichment = FALSE,
  organism = NULL,
  run_autocorrelation = FALSE,
  run_pixel_enrichment = FALSE,
  has_multi_modal = FALSE,
  multi_modal_path = NULL
)
}
\arguments{
\item{app_dir}{Directory to write the app.R file}

\item{bulk_intensity_matrix}{Path to or object for bulk intensity matrix}

\item{bulk_metadata}{Path to or object for bulk metadata}

\item{intensity_matrix}{Path to or object for intensity matrix}

\item{metadata}{Path to or object for metadata}

\item{anno}{Path to or object for annotation data}
}
\value{
Path to the written app.R file
}
\description{
This function writes a static app.R script using the provided data objects, matching the current example output of generateShinyApp.
}
