% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semla_spatial_autocorrelation.R,
%   R/semla_spatial_network.R
\name{CorSpatialFeatures.Seurat}
\alias{CorSpatialFeatures.Seurat}
\alias{CorSpatialFeatures}
\title{Find features with high spatial autocorrelation}
\usage{
\method{CorSpatialFeatures}{Seurat}(
  object,
  features = NULL,
  assay_use = NULL,
  slot_use = "data",
  across_all = FALSE,
  nCores = NULL,
  verbose = TRUE,
  ...
)

CorSpatialFeatures(object, ...)

CorSpatialFeatures(object, ...)
}
\arguments{
\item{object}{An object (see details)}

\item{features}{A character vector with features present in \code{Seurat} object. These
features need to be accessible with \code{\link{FetchData}}}

\item{assay_use}{Select assay to use for computation. If not specified, the default
assay will be used.}

\item{slot_use}{Select slot to use from assay object.}

\item{...}{Arguments passed to other methods}
}
\description{
Find features with high spatial autocorrelation

Find features with high spatial autocorrelation
}
\examples{

se_mbrain <- readRDS(system.file("extdata/mousebrain",
                                 "se_mbrain",
                                 package = "semla"))
se_mbrain <- se_mbrain |>
  ScaleData() |>
  RunPCA()

# Compute spatial autocorrelation for variable features
spatgenes <- CorSpatialFeatures(se_mbrain,
                                features = VariableFeatures(se_mbrain),
                                nCores = 1)

# Check genes with highest spatial autocorrelation
head(spatgenes[[1]])

# Note that the top variable genes are blood related (hemoglobin genes)
# These genes have lower spatial autocorrelation since blood vessels
# typically only cover a few spots and more randomly dispersed throughput the tissue
head(VariableFeatures(se_mbrain))

\donttest{
# The same principle can be used to estimate spatial autocorrelation for other features,
# for example dimensionality reduction vectors
spatpcs <- CorSpatialFeatures(se_mbrain,
                             features = paste0("PC_", 1:10),
                             nCores = 1)

# Calculate spatial autocorrelation scores for principal components
head(spatpcs[[1]])

# Compute spatial autocorrelation scores for multiple datasets
se_mcolon <- readRDS(system.file("extdata/mousecolon",
                                 "se_mcolon",
                                 package = "semla"))
se_merged <- MergeSTData(se_mbrain, se_mcolon) |>
  FindVariableFeatures()

spatgenes <- CorSpatialFeatures(se_merged,
                                features = VariableFeatures(se_merged),
                                nCores = 1)

# Check spatial autocorrelation scores mouse brain data
head(spatgenes[[1]])
# Check spatial autocorrelation scores mouse colon data
head(spatgenes[[2]])
}

}
\seealso{
Other spatial-methods: 
\code{\link{GetSpatialNetwork.default}()}

Other network-methods: 
\code{\link{GetSpatialNetwork.default}()}

Other network-methods: 
\code{\link{GetSpatialNetwork.default}()}
}
\author{
Ludvig Larsson
}
\concept{network-methods}
\concept{spatial-methods}
