% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPanel_functions.R
\name{create_bulk_exp}
\alias{create_bulk_exp}
\title{Create bulk dataset from pixel-wise intensity matrix}
\usage{
create_bulk_exp(
  expression.matrix,
  metadata,
  sample.id.column = 4,
  sample.wide.columns = c(),
  organism = "Mouse",
  adducts = c("M-H [1-]", "M-H20-H [1-]", "M+Cl [1-]"),
  mode = "Negative",
  ppm = 5
)
}
\arguments{
\item{expression.matrix}{A data.frame containing m/z values on the columns and pixels/spots on the rows.}

\item{metadata}{A data.frame with the first column matching the rownames from expression.matrix, the next 2
columns containing x and y coordinates and other columns describing attributes of the pixel, e.g. sample
of origin, timepoint, treatment, pathology annotations.}

\item{sample.id.column}{Name or index of column in metadata table which specifies the sample of origin}

\item{sample.wide.columns}{Name or indices of columns in metadata table other than sample.id.column which
contain sample-wide information, such as timepoint or treatment}
}
\value{
A numeric vector.
}
\description{
Create bulk dataset from pixel-wise intensity matrix
}
\examples{
add(1, 1)
add(10, 1)
}
