% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateShinyApp.R
\name{generateShinyApp}
\alias{generateShinyApp}
\title{Generate all files required for an autonomous MSI toolkit shiny app}
\usage{
generateShinyApp(
  shiny.dir = "MSIToolKitApp",
  intensity.matrix,
  metadata,
  sample.id.column,
  sample.wide.columns,
  only.annotated = TRUE,
  extra.annotation = NULL,
  organism = "Mouse",
  adducts = c("M-H [1-]", "M-H20-H [1-]", "M+Cl [1-]"),
  mode = "Negative",
  ppm = 5,
  run.svm = F,
  run.spe = F
)
}
\arguments{
\item{shiny.dir}{directory to store the shiny app}

\item{intensity.matrix}{the intensity matrix, a dataframe where rows correspond to
pixels (named by pixel ID) and columns correspond to peaks (named by m/z value).}

\item{metadata}{a data frame containing metadata for each pixel contained
in the intensity.matrix; must contain at minimum four columns:
the first column must contain the row names of the intensity matrix, the pixel IDs,
the second and third columns should be named x and y and contain (x,y) coordinates for each pixel,
and another column should correspond to the sample IDs which each pixel comes from}

\item{sample.id.column}{name of the column in metadata containing sample ID information where
each pixel comes from}

\item{sample.wide.columns}{name of any other columns which describe sample-wide information, e.g.
treatment, disease etc}

\item{only.annotated}{boolean argument describing whether only peaks which can be mapped to metabolites
should be included in the app (default: TRUE)}

\item{organism}{organism name to be used for metabolite mapping and pathway analysis.
Options are Human, Mouse or Rat (default: Mouse)}

\item{adducts}{Vector of adducts which should be used for metabolite mapping. Options are  M-H [1-],M-2H [2-],
M-3H [3-],M-H2O-H [1-],M-H+O [1-],M+K-2H [1-],M+Na-2H [1-],M+Cl [1-],M+Cl37 [1-],M+FA-H [1-],M+Hac-H [1-] ,
M+Br [1-],M+Br81 [1-],M+TFA-H [1-],M+ACN-H [1-],M+HCOO [1-],M+CH3COO [1-],2M-H [1-],2M+FA-H [1-],2M+Hac-H [1-],3M-H [1-],
M(C13)-H [1-],M(S34)-H [1-],M(Cl37)-H [1-] for negative mode and  M [1+],M+H [1+],M+2H [2+],M+3H [3+],M+Na [1+],M+2Na [2+]
M+3Na [3+],M+H+Na [2+],M+H+2Na [3+],M+2H+Na [3+],M+2Na-H [1+],M+NaCl [1+],M+K [1+],M+H+K [2+],M+ACN+H [1+],M+ACN+2H [2+],
M+ACN+Na [1+],M+2ACN+2H [2+],M+3ACN+2H [2+],M+2ACN+H [1+],M+H2O+H [1+],M-H2O+H [1+],M-H4O2+H [1+],M-HCOOH+H [1+],M+HCOONa [1+],
M-HCOONa+H [1+],M+HCOOK [1+],M-HCOOK+H [1+],M-CO+H [1+],M-CO2+H [1+],M-C3H4O2+H [1+],M+CH3OH+H [1+],M-NH3+H [1+],M+H+NH4 [2+],
M+NH4 [1+],M+IsoProp+H [1+],M+IsoProp+Na+H [1+],M+2K+H [1+],M+DMSO+H [1+],2M+H [1+],2M+NH4 [1+],2M+Na [1+],2M+3H2O+2H [2+],
2M+K [1+],2M+ACN+H [1+],2M+ACN+Na [1+],M(C13)+H [1+],M(C13)+2H [2+],M(C13)+3H [3+],M(S34)+H [1+],M(Cl37)+H [1+]  for positive mode}

\item{mode}{Whether the data comes from positive or negative mode (default: negative)}

\item{ppm}{ppm tolerance for metabolite mapping (default: 5)}
}
\description{
This function creates an app.R file and all required objects
to run the app in .rda format in the target directory. Basic checks are
performed to avoid problems with input data format. The app directory
is standalone and can be used on another platform, as long as the MSI toolkit
package is installed there.
}
