% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing_autocorrelation.R
\name{preprocessing_compute_autocorrelation}
\alias{preprocessing_compute_autocorrelation}
\title{Compute spatial autocorrelation (correlation between feature and spatial lag)}
\usage{
preprocessing_compute_autocorrelation(intensity, lag_mat)
}
\arguments{
\item{intensity}{Matrix (pixels x features).}

\item{lag_mat}{Matrix (pixels x features).}
}
\value{
Named vector of autocorrelation per feature.
}
\description{
Compute spatial autocorrelation (correlation between feature and spatial lag)
}
\keyword{internal}
