% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing_pixelenrichment.R
\name{preprocessing_get_differential_peaks}
\alias{preprocessing_get_differential_peaks}
\title{Get differential peaks for a pixel neighbourhood}
\usage{
preprocessing_get_differential_peaks(
  x_tf,
  y_tf,
  intensity.per.sample,
  metadata.per.sample,
  control.matrix,
  min.neighbours = 3,
  pvalue.cutoff = 0.05,
  lfc.cutoff = 0.5,
  test = "t-test",
  anno
)
}
\arguments{
\item{x_tf}{Numeric; x_tf-coordinate of the target spot.}

\item{y_tf}{Numeric; y_tf-coordinate of the target spot.}

\item{intensity.per.sample}{Matrix or data frame of intensities (spots x features) for the sample.}

\item{metadata.per.sample}{Data frame of spot metadata with x_tf, y_tf and spot_id columns.}

\item{control.matrix}{Matrix of control sample intensities (features x control spots).}

\item{min.neighbours}{Integer; minimum number of neighbouring spots required (default 3).}

\item{pvalue.cutoff}{Numeric; adjusted p-value cutoff for significance.}

\item{lfc.cutoff}{Numeric; log-fold-change cutoff for calling significance.}

\item{test}{Character; statistical test to use (default 't-test').}

\item{anno}{Data frame; annotation data for features used by \code{\link{bulk_utils_DA}}.}
}
\value{
Data frame of significantly differential peaks (or NULL if neighbourhood too small).
}
\description{
For a given coordinate (x_tf, y_tf), compute differential peaks between the local neighbourhood and control samples using \code{\link{bulk_utils_DA}}.
}
\keyword{internal}
