% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{preprocessing_get_gcds}
\alias{preprocessing_get_gcds}
\title{Compute grid spacing (GCD) for each sample}
\usage{
preprocessing_get_gcds(metadata)
}
\arguments{
\item{metadata}{Data frame with 'Sample', 'spot_id', 'x', and 'y' columns.}
}
\value{
Numeric; unique GCD value for the grid spacing.
}
\description{
Calculates the grid spacing (GCD) for each sample in the metadata, ensuring consistency across samples.
}
\keyword{internal}
